Fix all (non-deprecation) compiler warnings (#3321)
* Fix all (non-deprecation) compiler warnings

* satisfy spotless

* use a method ref over a lambda

+++ src/main/java/vazkii/botania/common/core/helper/ExperienceHelper.java
@@ -35,7 +35,7 @@ public class ExperienceHelper {
 		}
 
 		if (level > 0 && level < 17) {
-			return (int) (level * level + 6 * level);
+			return (level * level + 6 * level);
 		} else if (level > 16 && level < 32) {
 			return (int) (2.5 * level * level - 40.5 * level + 360);
 		} else {
